Assume process_function is callable
process_function is either a function with one argument (to be called
for each block), or zero arguments (to be spawned as an actor that is
polling for blocks to process). If it is an actor function, it should
use the environment variable DAISY_SCHED_ADDR_PORT, which is set by the
scheduler prior to calling the actor. It should not be a function that
returns a command; neither should it be assumed that it uses a
particular script like run_lsf or run_docker.