Fix sporadic hang bug when finishing tasks with local processes
The problem is that python sometime fails to terminate asyncio
threads of spawned processes, leading to main python process
hanging when finishing scheduling. This happens more often when
processes are rapidly spawn and terminated as often the case
for testcases.

Although not thoroughly confirmed, this is probably a problem with
implementation of asyncio. The fix however is to force terminate
(i.e., SIGTERM) all spawn processes when finishing scheduling.