Add multi-queue for DependencyGraph
Initially DependencyGraph.next() operated through a single queue
of block, and will only release the front of the queue to the
Scheduler. This means that if the Scheduler doesn't have an idle
worker for this block, it will stall even if it has workers for
the next item in the queue.

This commit tries to fix this problem by (1) splitting the single
job queue into multiple queues, one per task, and (2) allowing
the Scheduler to pass in availability of workers as a parameter.
The DependencyGraph first search for and release a block
with matching idle workers, and if not found, release any other
block for the Scheduler to wait on or spawn workers for.