fix: Shorten shutdown latency due to status thread
The status loop thread sleep period was set to 10s, which means that
because Daisy waits for the status loop, the shutdown/finishing latency
for Daisy was a long random period between 0 to 10s.

This fix reduces the thread sleep time to 0.1s, but keep the status line
print period at 10s.