Add `init_callback_fn` argument to `Task`
`init_callback_fn` is a function that Daisy will call once when the
task is started. It will be called as:

	init_callback_fn(context)

Where `context` is the `daisy.Context` string that can be used by
the daisy clients to connect to the server. This feature is useful
for scripts that need access to the context to generate submit
scripts and show to the user but does not necessarily has to run
for every worker.